Move _get_port_connected_node_map to port.py
_get_port_connected_node_map gives the node map containing ForceIP.
It sholdn't be a part of ImageAcquirer, since ForceIP should be usable
before connecting to the camera. It was moved to port.py since it's
connected to ports, however it could be moved elsewhere.
_get_port_connected_node_map is only slightly modified with arguments
being changed to function without being a part of a class.
